Updates: Correct documentation in classes file, Add the option to extract ns in the solver results and add ImmutableDenseMatrix to utils

+++ sympt/classes.py
@@ -1460,7 +1460,7 @@ class RDSymbol(Symbol):
 
     Attributes:
     -----------
-    order : int
+    order : float
         The order of the symbol (used in specific algebraic contexts).
 
     Methods:
